\name{aods3-pkg}
\alias{aods3-pkg}
\alias{aods3}
\docType{package}
\encoding{latin1}

\title{Analysis of Overdispersed Data}

\description{This package provides functions to analyse
        overdispersed counts or proportions. The functions should be considered as
        complements to more sophisticated methods such as generalized
        estimating equations (GEE) or generalized linear mixed effect
        models (GLMM).}

\details{

Functions Index :
\tabular{ll}{
  aodml       \tab ML Estimation of Generalized Linear Models for Overdispersed Count Data\cr
  aodql       \tab QL/MM Estimation of Generalized Linear Models for Overdispersed Count Data\cr
  drs         \tab Test of Proportion Homogeneity between Groups using Donner's and Rao-Scott's Adjustments\cr
  gof         \tab Test of Goodness-of-Fit of Models for Count data\cr
  iccbin      \tab Intra-Cluster Correlation for Clustered Binomial data\cr
  invlink     \tab Transformation from the Link Scale to the Observation Scale\cr
  link        \tab Transformation from the Observation Scale to the Link Scale\cr
  splitbin    \tab Split Grouped Data Into Individual Data\cr
  varbin      \tab Estimate of a Probability from Clustered Binomial Data\cr
  wald.test       \tab Wald Test for Model Coefficients\cr
  }
  
Data sets Index :
\tabular{ll}{
  antibio              \tab Antibiotics against Shipping Fever in Calves\cr
  cohorts              \tab Data set: Age, Period and Cohort Effects for Vital Rates\cr
  dja                  \tab Mortality of Djallonke Lambs in Senegal\cr
  lizards              \tab A Comparison of Site Preferences of Two Species of Lizard\cr
  mice                 \tab Pregnant Female Mice Experiment\cr
  orob1                \tab Germination Data\cr
  orob2                \tab Germination Data\cr
  rabbits              \tab Rabbits Foetuses Survival Experiment\cr
  rats                 \tab Rats Diet Experiment\cr
  salmonella           \tab Salmonella Reverse Mutagenicity Assay\cr
  }
  
}

\author{
  Matthieu Lesnoff <matthieu.lesnoff@cirad.fr> and Renaud Lancelot <renaud.lancelot@cirad.fr>\cr
  Maintainer: Renaud Lancelot <renaud.lancelot@cirad.fr>
  }
  
\keyword{package}
